+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 * Performance problem for `hexapdf inspect --pages` when inspecting huge files
 * TrueType compound glyph component offset calculation
+* Parsing of TrueType data type 'fixed'
 
 ## 0.2.0 - 2016-11-28